+++ source/winapi/trackbar.cpp
@@ -22,17 +22,18 @@
 
 HB_FUNC( HWG_INITTRACKBAR )
 {
-   auto hTrackBar = CreateWindowEx(0, TRACKBAR_CLASS,
-                                 0,
-                                 ( LONG ) hb_parnl(3),
-                                 hwg_par_int(4),
-                                 hwg_par_int(5),
-                                 hwg_par_int(6),
-                                 hwg_par_int(7),
-                                 hwg_par_HWND(1),
-                                 reinterpret_cast<HMENU>(static_cast<UINT_PTR>(hb_parni(2))),
-                                 GetModuleHandle(nullptr),
-                                 nullptr);
+   auto hTrackBar = CreateWindowEx(0,
+                                   TRACKBAR_CLASS,
+                                   0,
+                                   hb_parnl(3),
+                                   hwg_par_int(4),
+                                   hwg_par_int(5),
+                                   hwg_par_int(6),
+                                   hwg_par_int(7),
+                                   hwg_par_HWND(1),
+                                   reinterpret_cast<HMENU>(static_cast<UINT_PTR>(hb_parni(2))),
+                                   GetModuleHandle(nullptr),
+                                   nullptr);
 
    hb_retptr(hTrackBar);
 }
